kswapd measurements code

+++ rdmemusg.h
@@ -8,7 +8,7 @@ unsigned long getinfo(const char *file, const char *name);
 unsigned long meminfo(const char *name);
 unsigned long statusinfo(const char *name);
 unsigned long slabinfo(const char *name);
-unsigned long get_CPU_usage(int);
+double get_CPU_usage(int);
 unsigned long get_uptime();
 unsigned long get_uptime_jiffies();
